requirements.txt
streamlit==1.56.0
pandas==3.0.2
psycopg2-binary
plotly==6.7.0